#25 UPDATE sectors.py
- REARRANGE sectors.py and sector data
